# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  filesystem.Stat:
    properties:
      mimetype:
        type: string
      ufs.FileInfo: {}
    type: object
...
  installer.ServerDetails:
    properties:
      start_on_completion:
        type: boolean
      uuid:
        type: string
    type: object
...
  router.ErrorResponse:
    properties:
      error:
        type: string
      request_id:
        type: string
    type: object
...
  router.ServerArchiveRequest:
    properties:
      extension:
        type: string
      files:
        items:
          type: string
        type: array
      name:
        type: string
      root:
        type: string
    type: object
...
  router.ServerChmodRequest:
    properties:
      files:
        items:
          $ref: '#/definitions/router.ServerChmodFile'
        type: array
      root:
        type: string
    type: object
...
  router.ServerCopyRequest:
    properties:
      location:
        type: string
    type: object
  router.ServerCreateDirectoryRequest:
    properties:
      name:
        type: string
      path:
        type: string
    type: object
  router.ServerDecompressRequest:
    properties:
      file:
        type: string
      root:
        type: string
    type: object
...
  server.APIResponse:
    properties:
      configuration:
        $ref: '#/definitions/server.Configuration'
      is_suspended:
        type: boolean
      state:
        type: string
      utilization:
        $ref: '#/definitions/server.ResourceUsage'
    type: object
...
info:
  contact: {}
  description: API documentation for the FeatherWings daemon.
  title: FeatherWings API
  version: "1.0"
paths:
...
  /api/servers:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/server.APIResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      security:
      - NodeToken: []
      summary: List servers
      tags:
      - Servers
    post:
      consumes:
      - application/json
      parameters:
      - description: Server configuration
        in: body
        name: server
        required: true
        schema:
          $ref: '#/definitions/installer.ServerDetails'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      security:
      - NodeToken: []
      summary: Create server
      tags:
      - Servers
  /api/servers/{server}:
    delete:
      parameters:
      - description: Server identifier
        in: path
        name: server
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      security:
      - NodeToken: []
      summary: Delete server
      tags:
      - Servers
    get:
      parameters:
      - description: Server identifier
        in: path
        name: server
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      security:
      - NodeToken: []
      summary: Get server
      tags:
      - Servers
...
  /api/servers/{server}/files/chmod:
    post:
      consumes:
      - application/json
      parameters:
      - description: Server identifier
        in: path
        name: server
        required: true
        type: string
      - description: Chmod request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/router.ServerChmodRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      security:
      - NodeToken: []
      summary: Change file permissions
      tags:
      - Server Files
  /api/servers/{server}/files/compress:
    post:
      consumes:
      - application/json
      parameters:
      - description: Server identifier
        in: path
        name: server
        required: true
        type: string
      - description: Archive request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/router.ServerArchiveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/filesystem.Stat'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      security:
      - NodeToken: []
      summary: Compress files
      tags:
      - Server Files
  /api/servers/{server}/files/contents:
    get:
      parameters:
      - description: Server identifier
        in: path
        name: server
        required: true
        type: string
      - description: File path
        in: query
        name: file
        required: true
        type: string
      - description: Force download
        in: query
        name: download
        type: boolean
      produces:
      - text/plain
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      security:
      - NodeToken: []
      summary: Read file contents
      tags:
      - Server Files
  /api/servers/{server}/files/copy:
    post:
      consumes:
      - application/json
      parameters:
      - description: Server identifier
        in: path
        name: server
        required: true
        type: string
      - description: Copy request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/router.ServerCopyRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      security:
      - NodeToken: []
      summary: Copy file
      tags:
      - Server Files
  /api/servers/{server}/files/create-directory:
    post:
      consumes:
      - application/json
      parameters:
      - description: Server identifier
        in: path
        name: server
        required: true
        type: string
      - description: Directory request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/router.ServerCreateDirectoryRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      security:
      - NodeToken: []
      summary: Create directory
      tags:
      - Server Files
  /api/servers/{server}/files/decompress:
    post:
      consumes:
      - application/json
      parameters:
      - description: Server identifier
        in: path
        name: server
        required: true
        type: string
      - description: Decompression request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/router.ServerDecompressRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      security:
      - NodeToken: []
      summary: Decompress archive
      tags:
      - Server Files
  /api/servers/{server}/files/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: Server identifier
        in: path
        name: server
        required: true
        type: string
      - description: Delete request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/router.ServerDeleteRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      security:
      - NodeToken: []
      summary: Delete files
      tags:
      - Server Files
  /api/servers/{server}/files/list-directory:
    get:
      parameters:
      - description: Server identifier
        in: path
        name: server
        required: true
        type: string
      - description: Directory path
        in: query
        name: directory
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/filesystem.Stat'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      security:
      - NodeToken: []
      summary: List directory contents
      tags:
      - Server Files
...
  /api/servers/{server}/files/pull/{download}:
    delete:
      parameters:
      - description: Server identifier
        in: path
        name: server
        required: true
        type: string
      - description: Download identifier
        in: path
        name: download
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - NodeToken: []
      summary: Cancel remote download
      tags:
      - Server Files
...
schemes:
- https
- http
securityDefinitions:
  NodeToken:
    description: 'Supply the node''s bearer token from `config.yml` using the `Authorization:
      Bearer <token>` header.'
    in: header
    name: Authorization
    type: apiKey
  ServerJWT:
    description: Signed JWTs issued by the Panel for server-scoped operations (uploads,
      downloads, websockets). Pass in the `token` query parameter.
    in: query
    name: token
    type: apiKey
swagger: "2.0"